Delete trip dashboard from among markdown file
When the code is run, in R Markdown, a file containing all the created images is made. I want to create a separate folder in which I upload images manually so that upon running the code, updates will only be made to images that are part of the code as opposed to uploaded files which in turn will be lost due to updates.